Use progress bar

+++ Programm/Settings.hpp
@@ -91,7 +91,7 @@ class Settings {
           */
         int accept_seed {0};
 
-        int bootstrap_samples {1000};
+        size_t bootstrap_samples {1000};
 
         /**
           Number of bins for the position histogram.
